Demo_and_Hosting/requirements.txt
streamlit
pandas>=2.0
numpy>=1.24
plotly
scikit-learn==1.6.1
joblib>=1.3
langgraph>=0.2
langchain>=0.3
langchain-text-splitters
langchain-openai
langchain-community
langchain-huggingface
chromadb
protobuf>=3.20.0,<4.0.0
sentence-transformers
python-dotenv
requests
